Refactor Crypto
