+++ core/src/main/java/org/apache/accumulo/core/conf/Property.java
@@ -176,7 +176,7 @@ public enum Property {
       "Sensitive properties related to on-disk file encryption."),
   @Experimental
   INSTANCE_CRYPTO_SERVICE("instance.crypto.service",
-      "org.apache.accumulo.core.cryptoImpl.NoCryptoService", PropertyType.CLASSNAME,
+      "org.apache.accumulo.core.spi.crypto.NoCryptoService", PropertyType.CLASSNAME,
       "The class which executes on-disk file encryption. The default does nothing. To enable "
           + "encryption, replace this classname with an implementation of the"
           + "org.apache.accumulo.core.spi.crypto.CryptoService interface."),
