+++ core/src/test/resources/accumulo2.properties
@@ -18,7 +18,7 @@
 #
 
 general.rpc.timeout=123s
-instance.crypto.service=org.apache.accumulo.core.cryptoImpl.AESCryptoService
+instance.crypto.service=org.apache.accumulo.core.spi.crypto.AESCryptoService
 instance.secret=mysecret
 instance.volumes=hdfs://localhost:8020/accumulo123
 instance.zookeeper.host=myhost123:2181
